Big refactor in WaveletFilterBank.
Rename: Forward / InverseWaveletFilterBankFFT to
WaveletFrequencyFilterBankForwardGenerator

FilterBanks now derives from GenerateImageSource, instead of from
ImageToImageFilter. So no Input is required.

Forward/Inverse FilterBank are exactly the same, except that the
first calls EvaluateForwardSubBand from the Evaluator (a IsotropicWavelet)
and the second calls EvaluateInverseSubBand instead . In the
case of Held Wavelet, the output is the same --it just felt weird to
have a m_InverseWaveletFlag that has to be checked every time... The
really bad thing is the code duplication for just one line.
What about the template bool flag, and an EvaluateXXX specialization, OTB style?
I don't like the template instantiation and the need of a .cpp
(library). What about inline?
Puff, partial template specialization of a class is painful in c++03.
In c++11 you can use std::enable_if, or std::is_same, and done!
I would like to stick to pre c++11 just in case I submit it to ITK for
merge.