Merge headers of Forward/Inverse filter bank.
Same filter, with a SetInverseBank(true/false)

The good think is no code-duplication.

The bad thing about this approach, is that user can be tempted to
first calculate forward filters, and then using the same filter,
calculate inverse. This will trigger modified() in the already
calculated forward, screwing up whatever is connected to the forward
filter in the pipeline....

Partial template specialization should be a thing, but without c++11
seems messy...